Bestscore = score

+++ ggp-base-master/src/org/ggp/base/player/gamer/statemachine/GameNode.java
@@ -10,7 +10,7 @@ public class GameNode {
 	GameNode parent;
 	MachineState state;
 	int visits;
-	int utility;
+	double utility;
 	List<GameNode> children;
 	Move move;
 	public GameNode(MachineState state, GameNode parent, Move move)
